chore(provider/minimax): split quota tests into sibling file (CI: file-size)
The security-hardening additions pushed `quota.rs` to 403 lines,
3 over the 400-line cap. CI's File Size Lint failed on PR #843
while local runs pre-hardening were clean. Extracts the inline
`mod tests` block into `quota_tests.rs` via the `#[cfg(test)]
#[path = ...] mod tests;` pattern already used for `parser_tests.rs`,
`tests.rs`, `sse_tests.rs`, and `cost_tests.rs` in this PR. No
behavior change; all 9 quota tests still pass.
